+++ CHANGELOG.md
@@ -10,12 +10,12 @@
   - Add vertical swipe navigation for scrolling of menu items.
   - Refactored options screen to use lv_list instead of a custom flex container to improve performance
   - Replaced single lv_textarea in terminal view with scrollable lv_page and per-line lv_label children to improve performance
-
+  - Offset terminal page vertically by status bar height and adjust its height accordingly.
+  
 ### Bug Fixes
 
 - Display
   - Fixed an issue where an option would be duplicated and freeze the device.
-